Changed fibonacci function to use binary power and matrix to faster computations and allow to use bigger N
